mithril-backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Task queue
taskiq>=0.11.0
taskiq-redis>=1.0.0
redis>=5.0.0

# Firebase
firebase-admin>=6.5.0

# AWS
boto3>=1.35.0

# Google AI
google-genai>=1.0.0

# Google Cloud Monitoring
google-cloud-monitoring>=2.21.0

# Image processing
pillow>=10.4.0

# HTTP client
httpx>=0.27.0

# xAI SDK
xai-sdk>=0.0.1

# 3D rendering
numba>=0.60.0              # JIT-accelerated tile-based Gaussian splatting
trimesh>=4.0.0
plyfile>=0.8.1          # 3DGS .ply loading
gaussforge             # 3DGS format conversion (SPZ, SPLAT, KSPLAT, SOG → PLY)
PyOpenGL>=3.1.7
networkx>=2.2
freetype-py>=2.2.0
imageio>=2.14.1
pyglet>=2.0.0
# pyrender installed via Dockerfile with --no-deps to avoid PyOpenGL downgrade

# Settings management
pydantic-settings>=2.5.0